+++ packages/admin-api-client/src/methods.ts
@@ -71,35 +71,3 @@ export class AdminApiClient {
     return this._adminApiClient.configClient(config);
   }
 }
-
-const defaultClient = new AdminApiClient();
-
-export const readAbout = defaultClient.readAbout.bind(defaultClient);
-
-export const readConfig = defaultClient.readConfig.bind(defaultClient);
-
-export const updateConfig = defaultClient.updateConfig.bind(defaultClient);
-
-export const readAlerts = defaultClient.readAlerts.bind(defaultClient);
-
-export const readAlert = defaultClient.readAlert.bind(defaultClient);
-
-export const readCollections = defaultClient.readCollections.bind(defaultClient);
-
-export const readCollection = defaultClient.readCollection.bind(defaultClient);
-
-export const readRoutes = defaultClient.readRoutes.bind(defaultClient);
-
-export const readRoute = defaultClient.readRoute.bind(defaultClient);
-
-export const readVariants = defaultClient.readVariants.bind(defaultClient);
-
-export const readVariant = defaultClient.readVariant.bind(defaultClient);
-
-export const readCustomRouteVariants = defaultClient.readCustomRouteVariants.bind(defaultClient);
-
-export const useRouteVariant = defaultClient.useRouteVariant.bind(defaultClient);
-
-export const restoreRouteVariants = defaultClient.restoreRouteVariants.bind(defaultClient);
-
-export const configClient = defaultClient.configClient.bind(defaultClient);