+++ packages/admin-api-paths/src/index.ts
@@ -1,13 +1,3 @@
-export const LEGACY_DEFAULT_BASE_PATH = "/admin";
-export const LEGACY_SETTINGS = "/settings";
-export const LEGACY_ABOUT = "/about";
-export const LEGACY_ALERTS = "/alerts";
-
-export const LEGACY_MOCKS = "/mocks";
-export const LEGACY_ROUTES = "/routes";
-export const LEGACY_ROUTES_VARIANTS = "/routes-variants";
-export const LEGACY_MOCK_CUSTOM_ROUTES_VARIANTS = "/mock-custom-routes-variants";
-
 export const BASE_PATH = "/api";
 export const CONFIG = "/config";
 export const ABOUT = "/about";