+++ packages/core/src/common/events.js
@@ -19,6 +19,5 @@ function addEventListener(listener, eventName, eventEmitter) {
 module.exports = {
   CHANGE_MOCK: "change:mock",
   CHANGE_ALERTS: "change:alerts",
-  CHANGE_LOGS: "change:logs",
   addEventListener,
 };