+++ packages/core/src/scaffold/Scaffold.js
@@ -150,8 +150,9 @@ class Scaffold {
 
     if (this._readConfigFileOption.value && !configFileLoaded) {
       this._alerts.set("config", "Configuration file was not found. A scaffold was created");
-      // LEGACY, Set base collection, which is the one created in the scaffold, when selected collection is not defined in the option. (when legacy option is removed)
-      // If the collection selected option value is not set, it should be set as "base"
+      if (!this._collectionSelectedOption.hasBeenSet) {
+        this._collectionSelectedOption.value = "base";
+      }
       return this._createConfig();
     }
     return Promise.resolve();