+++ packages/plugin-admin-api/src/Alerts.js
@@ -31,7 +31,6 @@ class AlertsApi {
       getItems: this._getCollection.bind(this),
       parseItem: this._parseModel.bind(this),
       logger: this._logger.namespace("alert"),
-      finder: (context) => (item) => item.context === context,
     });
   }
 
@@ -40,11 +39,11 @@ class AlertsApi {
   }
 
   _parseCollection() {
-    return this._alerts.root.customFlat.map(this._parseModel.bind(this));
+    return this._alerts.root.flat.map(this._parseModel.bind(this));
   }
 
   _getCollection() {
-    return this._alerts.root.customFlat;
+    return this._alerts.root.flat;
   }
 
   get router() {