+++ test/core-e2e-legacy/src/fixtures/typescript-imports/routes/books.ts
@@ -1,3 +1,0 @@
-import books from "../../typescript/routes/books.json";
-
-export default books;