+++ test/core-e2e-legacy/src/fixtures/typescript-imports/routes/middlewares.ts
@@ -1,3 +1,0 @@
-import middlewares from "../../typescript/routes/middlewares";
-
-export default middlewares;