+++ test/core-e2e/src/fixtures/cjs-config/mocks.config.js
@@ -1,0 +1,17 @@
+const path = require("path");
+
+module.exports = {
+  log: "silly",
+  mock: {
+    collections: {
+      selected: "user-real",
+    }
+  },
+  files: {
+    babelRegister: {
+      enabled: false,
+    },
+    path: path.resolve(__dirname, "..", "temp"),
+    watch: false,
+  },
+};