+++ test/cypress-commands-typescript-e2e/mocks.config.js
@@ -6,7 +6,9 @@ module.exports = {
   config: {
     allowUnknownArguments: true,
   },
-  mocks: {
-    selected: "standard",
+  mock: {
+    collections: {
+      selected: "standard",
+    },
   },
 };