+++ packages/main/CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Breaking change
 
+## [unreleased]
+
+### Fixed
+- fix: Update @mocks-server/plugin-admin-api
+- fix: Update @mocks-server/core
+
 ## [3.12.0] - 2022-09-01
 
 ### Added