+++ packages/plugin-admin-api/CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [unreleased]
+
+### Removed
+- feat: Remove custom router "/admin" from mock server. Admin API is only available at its own server from now.
+
+### Added
+- feat: Add url and method to route not found log
+
 ## [3.4.0] - 2022-08-11
 
 ### Added